requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dazer==0.1.13
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
joblib==1.3.2
kombu==5.3.3
Markdown==3.5
numpy==1.26.1
pandas==2.1.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.9
gunicorn==21.2.0
rq==1.15.1
psycopg2-binary==2.9.9
